상품 가격 데이터 타입 double 으로 변경

+++ src/main/java/com/ming/mingcommerce/product/ProductCrawler.java
@@ -62,7 +62,7 @@ public class ProductCrawler {
             // 가격의 데이터타입을 string 에서 float 으로 변환한다.
             try {
                 price = price.replace("$", "");
-                product.setPrice(Float.parseFloat(price));
+                product.setPrice(Double.parseDouble(price));
             } catch (Exception e) { // 가격에 할인된 가격표시로 화살표 문자열이 포함되어있는 경우가 있어 파싱에러를 잡아 가격을 null 로 세팅.
                 product.setPrice(null);
             }
